+++ R/clean-contents.R
@@ -10,7 +10,7 @@
 #' @noRd
 clean_contents <- function(
     ht,
-    output_type = c("latex", "html", "screen", "markdown", "word", "excel", "rtf"),
+    output_type = c("latex", "html", "screen", "markdown", "word", "excel", "rtf", "typst"),
     ...) {
   output_type <- match.arg(output_type)
   contents <- as.matrix(as.data.frame(ht))
@@ -87,9 +87,13 @@ render_markdown_matrix <- function(contents, ht, output_type) {
     column <- contents[, col]
     md_rows <- md[, col]
     column[md_rows] <- render_markdown(column[md_rows], output_type)
-    if (output_type %in% c("latex", "html", "rtf")) {
+    if (output_type %in% c("latex", "html", "rtf", "typst")) {
       to_esc <- esc[, col] & !md_rows
-      column[to_esc] <- sanitize(column[to_esc], output_type)
+      column[to_esc] <- if (output_type == "typst") {
+        typst_escape(column[to_esc])
+      } else {
+        sanitize(column[to_esc], output_type)
+      }
     }
     column
   }, character(nrow(contents)))